nadia_guerouaou.md
I am a doctoral student in the FEMTO Neuro Team , the Perception and Sounds Design (STMS lab/CNRS/Ircam Paris) and Plasticity and Subjectivity (Lille Neuroscience & Cognition Centre lab/INSERM/CHRU Lille) teams. In parallel to my research activity, I have a degree in Neuropsychology and work as a clinical psychologist at the Regional Consultation of Psychotrauma in the Hauts de France, where I receive patients suffering from Post Traumatic Stress Disorder (PTSD).

My doctoral research work focuses on the voice of PTSD's patients. The first objective is to investigate new acoustic biomarkers in the patients' voice in order to better characterize both the pathology and the healing process associated with imaginative exposure psychotherapy. The second aspect of this doctoral project is the use of voice transformation techniques developed at IRCAM for the purpose of augmented psychotherapy.  

Besides, I am also interested in Neuroethics, a discipline I teach Psychology and Neuropsychology students (Bachelor and Master) at Lille’s University. The core of these lectures mainly deal with the current and potential uses of new technologies emerging from research in cognitive neurosciences and the impact that misuses, particularly in the context of augmentation, could have on our societies. These rapid technological advances, require continuous adaptation from users and experts, and open debates to overcome the techclash and bring out the essential questions for an enlightened use of these technologies.This interest is developed in my thesis work by the study of the acceptability of vocal deepfakes by the society, and in particular of expressive voice transformation technologies as well as the impact of culture on this acceptability (forthcoming study: comparison of the receptivity of vocal deepfakes between the French and Japanese populations).

Thesis co-supervised by Pr Guillaume Vaiva (Professor of Psychiatry - Lille's Hospital) and Dr Jean-Julien Aucouturier (FEMTO Neuro team)